Adds feature extension abstraction.
This makes it possible to separate constructors from the actions that most classes need to run in a consistent manner.

This also added some logic that forces the loader registry to run this function if the class happens to extend it.

The most powerful benefit of this, is it makes it possible to pass anonymous class to a loader registry instead of creating a file for EVERYTHING. This may make sense in some circumstances (like shortcodes) where there's just not a lot of stuff going on in the class.